OEM-EV-Warranty-Management-System/Vehicle-Management-Service/vms-backend/requirements.txt
Django
djangorestframework 
mysqlclient 
drf-yasg 
pika